+++ src/dawn/node/binding/Converter.h
@@ -87,7 +87,7 @@ class Converter {
     // The pointer assigned to 'out_els' is valid until the Converter is destructed.
     template <typename OUT, typename IN>
     [[nodiscard]] inline bool operator()(OUT*& out_els,
-                                         uint32_t& out_count,
+                                         size_t& out_count,
                                          const std::vector<IN>& in) {
         return Convert(out_els, out_count, in);
     }
@@ -357,7 +357,7 @@ class Converter {
 
     // vector -> raw pointer + count
     template <typename OUT, typename IN>
-    inline bool Convert(OUT*& out_els, uint32_t& out_count, const std::vector<IN>& in) {
+    inline bool Convert(OUT*& out_els, size_t& out_count, const std::vector<IN>& in) {
         if (in.size() == 0) {
             out_els = nullptr;
             out_count = 0;
@@ -376,7 +376,7 @@ class Converter {
     // unordered_map -> raw pointer + count
     template <typename OUT, typename IN_KEY, typename IN_VALUE>
     inline bool Convert(OUT*& out_els,
-                        uint32_t& out_count,
+                        size_t& out_count,
                         const std::unordered_map<IN_KEY, IN_VALUE>& in) {
         if (in.size() == 0) {
             out_els = nullptr;
@@ -396,7 +396,7 @@ class Converter {
 
     // std::optional<T> -> raw pointer + count
     template <typename OUT, typename IN>
-    inline bool Convert(OUT*& out_els, uint32_t& out_count, const std::optional<IN>& in) {
+    inline bool Convert(OUT*& out_els, size_t& out_count, const std::optional<IN>& in) {
         if (!in.has_value()) {
             out_els = nullptr;
             out_count = 0;
